[IMP] hr_attendance_reason: Move all related code to the same file
