+++ hr_attendance_reason/models/__init__.py
@@ -1,6 +1,5 @@
 from . import hr_attendance_reason
 from . import hr_attendance
 from . import hr_employee_base
-from . import hr_employee
 from . import res_company
 from . import res_config_settings
